Changed names of new plugins for dynamic masking.
